+++ gtests/dynamic/src/test_ilogb.c
@@ -7,7 +7,7 @@ int test_ilogb(void* handle) {
     func  s1d = (func)dlsym(handle, "amd_ilogb");
 
     /*scalar inputs*/
-    float inputf = 3.145, outputf;
+    float inputf = 3.145f, outputf;
     double input = 6.287, output;
 
     error = dlerror();
@@ -19,9 +19,9 @@ int test_ilogb(void* handle) {
     printf("Exercising ilogb routines\n");
     /*scalar*/
     outputf = s1f(inputf);
-    printf("amd_ilogbf(%f) = %f\n", inputf, outputf);
+    printf("amd_ilogbf(%f) = %f\n", (double)inputf, (double)outputf);
     output = s1d(input);
-    printf("amd_ilogb(%lf) = %lf\n", input, output);
+    printf("amd_ilogb(%lf) = %lf\n", (double)input, (double)output);
 
     return 0;
 }
